requirements.txt
pillow>=10.0.0
fpdf2>=2.7.0
pandas>=2.0.0
openpyxl>=3.1.0
bcrypt>=4.0.0